pyproject.toml
[project]
name = "auth-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Ziad Alnajjar",email = "ziad.alnajjar@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard] (>=0.119.0,<0.120.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic[email] (>=2.12.3,<3.0.0)",
    "alembic (>=1.17.0,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-ulid (>=3.1.0,<4.0.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "redis[hiredis] (>=6.4.0,<7.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-asyncio (>=1.2.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]
